python/requirements.txt
kafka-python
bs4
requests
lxml
websocket-client
